+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog Format
 - [Initials of Dev] [#issue number](issue link on GitHub) {issue title}
 -->
 
+- [KF] [#85](https://github.com/Ateneo-MISA/MISA-Website/issues/85) vote(feature): add coming soon text in button
 - [KF MF] [#82](https://github.com/ateneomisa/MISA-Website/issues/82) vote(feature): fix ux/ui bugs 230326
 - [KF MF] [#80](https://github.com/ateneomisa/MISA-Website/issues/80) vote(feature): add voting
 - [KF] [#76](https://github.com/Ateneo-MISA/MISA-Website/issues/76) vote(feature): fix hero text